# docs/swagger.yaml
basePath: /api
definitions:
  model.User:
    properties:
      createBy:
        description: 创建人
        type: integer
      createTime:
        description: 创建时间
        type: string
      deleted:
        description: 是否删除
        type: integer
      desc:
        description: 描述
        type: string
      email:
        description: 邮箱
        type: string
      face:
        description: 头像
        type: string
      id:
        description: id
        type: integer
      mobile:
        description: 手机号
        type: string
      name:
        description: 姓名
        type: string
      password:
        description: 密码
        type: string
      sex:
        description: 性别
        maximum: 2
        minimum: 1
        type: integer
      updateBy:
        description: 更新人
        type: integer
      updateTime:
        description: 更新时间
        type: string
    required:
    - email
    - mobile
    - name
    type: object
  response.Response:
    properties:
      code:
        description: 响应码
        type: integer
      data:
        description: 响应数据
      message:
        description: 响应信息
        type: string
    type: object
host: localhost:80
info:
  contact:
    email: 2366850717@qq.com
  description: 中药相关接口，包括用户、banner、中药、药方等
  title: 中药 API
  version: 1.0.0
paths:
  /banner/list:
    get:
      description: 查询全部banner
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 查询banner列表
  /chineseHerb/get/:id:
    get:
      description: 通过id查找中药详情
      parameters:
      - description: 中药id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 查找中药
  /chineseHerb/list:
    get:
      description: 通过条件查询中药列表
      parameters:
      - description: 中药名
        in: query
        name: q
        type: string
      - description: 中药类型
        in: query
        name: type
        type: integer
      - description: 分页大小
        in: query
        name: pageSize
        type: integer
      - description: 页码
        in: query
        name: pageNo
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 查询中药列表
  /upload/img:
    post:
      description: 图片上传
      parameters:
      - collectionFormat: csv
        description: 上传的图片
        in: formData
        items:
          type: integer
        name: file
        required: true
        type: array
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 图片上传
  /user/add:
    get:
      description: 新增用户
      parameters:
      - description: 请求body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.User'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 新增用户
  /user/get/:id:
    get:
      description: 通过id查找用户详情
      parameters:
      - description: 用户id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 查找用户
  /user/list:
    get:
      description: 通过条件查询用户列表
      parameters:
      - description: 用户名或手机号
        in: query
        name: q
        required: true
        type: string
      - description: 分页大小
        in: query
        name: pageSize
        type: integer
      - description: 页码
        in: query
        name: pageNo
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 查询用户列表
swagger: "2.0"